Split PopupDialog styles
